requirements.txt
python-telegram-bot==20.7
mcstatus==11.1.1
